alternating optimisation
